[BUGFIX] Tracked Props - Prevents autotracking ArrayProxy creation
This PR prevents an issue with inifinite rerendering that is
fundamentally caused by creating an array proxy and returning it from
a getter. This is commonly done with _promise arrays_, which proxy to
content that is fulfilled at a later point in time:

```js
export default CountrySelectComponent extends Component {
  @service store;

  get options() {
    return this.store.findAll('country');
  }
}
```

While it likely makes sense to memoize these types of requests, even in
that case, the cache will be busted by autotracking because the
_creation_ of the array proxy causes it's content to be tracked. When
the content is updated later, it will then invalidate the cache, causing
the value to be thrown away before it can even be used, and a new value
to be fetched. This loops infinitely.

To get around this, we realized that array proxies do not need to add
observers until they have been _interacted_ with in some way. We set the
`CONSUMED` flag on the proxy to indicate that this has happened, and
only add/remove observers if it has. Since this happens lazily, it is
_generally_ guaranteed to happen after the getter has returned, since it
usually doesn't make sense to both create a value and access it in the
same getter.